+++ packages/vector_graphics/CHANGELOG.md
@@ -15,6 +15,7 @@
 * Adds SVG `feMorphology` support with parameter, pixel, and visual reference tests.
 * Adds SVG `feComponentTransfer` support with parameter, pixel, and visual reference tests.
 * Adds SVG `feTile` support with parameter, pixel, and visual reference tests.
+* Adds SVG `feConvolveMatrix` support with parameter, pixel, and visual reference tests.
 
 ## 1.2.3